requirements.txt
streamlit
faiss-cpu
sentence-transformers
openai